+++ src/js/controls/animation-display.js
@@ -8,7 +8,7 @@ import videojs from 'video.js';
 const Component = videojs.getComponent('Component');
 
 /**
- * Image for displaying animated GIF image.
+ * Component for displaying animated GIF images.
  *
  * @class
  * @augments videojs.Component
@@ -26,6 +26,17 @@ class AnimationDisplay extends Component {
             className: 'vjs-animation-display',
             innerHTML: '<img />'
         });
+    }
+
+    /**
+     * Set `src` of image element.
+     *
+     * @param {object} data - TODO
+     */
+    load(data) {
+        let img = this.el().firstChild;
+
+        img.src = data;
     }
 }